backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pinecone-client==3.0.0
google-generativeai==0.3.2
tiktoken==0.5.2
python-dotenv==1.0.0
pydantic==2.5.3
aiofiles==23.2.1
pypdf==3.17.4
slowapi==0.1.9
